requirements.txt
requests_html
pyyaml
lxml_html_clean
